feat: refactor pagination types to use uint64 and update attribute handling in product attributes

+++ internal/delivery/http/request/manufacturer_request/paginate.go
@@ -1,6 +1,6 @@
 package manufacturer_request
 
 type GetManufacturerWithPagination struct {
-	Page     *uint32 `json:"page" query:"page"`
-	PageSize *uint32 `json:"page_size" query:"page_size"`
+	Page     *uint64 `json:"page" query:"page"`
+	PageSize *uint64 `json:"page_size" query:"page_size"`
 } // @name GetProductWithPaginationRequest
