+++ internal/dto/common.go
@@ -1,6 +1,6 @@
 package dto
 
 type GetDTO struct {
-	Page     *uint32 `json:"page" query:"page"`
-	PageSize *uint32 `json:"page_size" query:"page_size"`
+	Page     *uint64 `json:"page" query:"page"`
+	PageSize *uint64 `json:"page_size" query:"page_size"`
 }
